+++ CHANGELOG.md
@@ -27,6 +27,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased] - 2023-01-04
 
 ### ADDED
+- SerialBuffering fuction.
+
+---
+
+## [Unreleased] - 2023-01-04
+
+### ADDED
 - SendTempSerial fuction.
 - monitor_speed.